+++ nix/default.nix
@@ -11,7 +11,7 @@
 
 mkDerivation {
   pname = "instant-hashable";
-  version = "0.1.0.1";
+  version = "0.2";
   homepage = "https://github.com/k0001/instant-hashable";
   description = "Generic Hashable instances through instant-generics";
   license = stdenv.lib.licenses.bsd3;