+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# Version 0.2
+
+* Rename 'ghashWithSalt' to 'ghashWithSaltDefault'.
+
+* Export 'RepGHashable'
+
+* Export 'GHashable' internals.
+
+
 # Version 0.1.0.1
 
 * Raise upper bound on `instant-generics` dependency.